+++ test/logging.spec.js
@@ -20,8 +20,81 @@ const T = require('../index.js')
  */
 const DELAY = 800
 
+/**
+ * The expected error message whenever the log file path is malformed or cannot
+ * be accessed.
+ * @type {String}
+ */
+const ERR_INVALID_LOG_PATH = 'If specified, the "saveLogTo" option must refer to a valid location that this proces has write-access to.'
+
 describe('the logging functionality', () => {
-  context('when given a valid path spec for the log file', () => {
+  context('when the value of `saveLogTo` is not specified', () => {
+    before(() => {
+      return rmrf(path.join(__dirname, '*.log'))
+    })
+
+    after(() => {
+      return rmrf(path.join(__dirname, '*.log'))
+    })
+
+    it('must be created in the same location as the test file', () => {
+      const instance = new T({
+        command: global.scriptCommands.runsNormally,
+        waitFor: /ready/
+      })
+
+      return instance
+        .start()
+        .then(() => {
+          return new Promise((resolve) => {
+            global.setTimeout(resolve, DELAY) // wait for process to exit on its own
+          })
+        })
+        .then(() => {
+          return fs.readdir(__dirname)
+        })
+        .then((filenames) => {
+          const actual = filenames.find((name) => /.+\.log$/.test(name))
+          expect(actual).to.be.a('string')
+        })
+    })
+  })
+
+  context.skip('when the value of `saveLogTo` is a valid file spec', () => {})
+
+  context.skip('when the value of `saveLogTo` is `null`', () => {})
+
+  context('when the value of `saveLogTo` is not a string', () => {
+    describe('the constructor', () => {
+      it('must throw an error', () => {
+        expect(() => {
+          const instance = new T({ // eslint-disable-line no-unused-vars
+            waitFor: /something/,
+            saveLogTo: [404]
+          })
+        }).to.throw(ERR_INVALID_LOG_PATH)
+      })
+      /* eslint-enable no-unused-vars */
+    })
+  })
+
+  context('when the value of `saveLogTo` is a string, but does not point to a valid location', () => {
+    describe('the `start` method', () => {
+      it('must be rejected', () => {
+        const instance = new T({
+          command: global.scriptCommands.runsNormally,
+          waitFor: /ready/,
+          saveLogTo: '/this/path/does/not/exist/log.txt'
+        })
+
+        const promise = instance.start()
+
+        expect(promise).to.be.rejectedWith(ERR_INVALID_LOG_PATH)
+      })
+    })
+  })
+
+  context.skip('when given a valid path spec for the log file', () => {
     context('without a value for "reference"', () => {
       /**
        * An instance of the code under test.
@@ -142,7 +215,7 @@ describe('the logging functionality', () => {
         })
 
         it('must end with the exit code from the child process', () => {
-          const pattern = /^exit code: 0$/i
+          const pattern = /^exit code:\s+0$/i
           const actual = logFile.lines[logFile.lines.length - 1]
 
           expect(actual).to.match(pattern)
@@ -314,7 +387,96 @@ describe('the logging functionality', () => {
         })
 
         it('must end with the exit code from the child process', () => {
-          const pattern = /^exit code: 1$/i
+          const pattern = /^exit code:\s+1$/i
+          const actual = logFile.lines[logFile.lines.length - 1]
+
+          expect(actual).to.match(pattern)
+        })
+      })
+    })
+
+    context('when the process is forced to exit', () => {
+      /**
+       * An instance of the code under test.
+       * @type {Object}
+       */
+      let instance = null
+
+      /**
+       * The folder that the log file will be created in.
+       * @type {String}
+       */
+      let tempFolder = null
+
+      before(() => {
+        return fs.mkdtemp(path.join(os.tmpdir(), 'procmonrest-'))
+          .then((pathspec) => {
+            tempFolder = pathspec
+
+            instance = new T({
+              command: global.scriptCommands.doesNotExit,
+              waitFor: /ready/,
+              saveLogTo: path.join(tempFolder, 'test.log')
+            })
+
+            return instance
+              .start()
+              .then(() => {
+                return new Promise((resolve) => {
+                  global.setTimeout(resolve, 800)
+                })
+              })
+              .then(() => {
+                return instance.stop()
+              })
+              .then(() => {
+                return new Promise((resolve) => {
+                  global.setTimeout(resolve, 800) // this is needed for the final write to the log file
+                })
+              })
+          })
+      })
+
+      after(() => {
+        return rmrf(tempFolder, { force: true })
+      })
+
+      describe('the created file', () => {
+        /**
+         * Information about the created log file.
+         * @type {Object}
+         */
+        const logFile = {
+          /**
+           * The name (with extension) of the log file.
+           * @type {String}
+           */
+          name: '',
+
+          /**
+           * The contents of the file, separated by newline character.
+           * @type {Array}
+           */
+          lines: []
+        }
+
+        before(() => {
+          return fs.readdir(tempFolder)
+            .then((list) => {
+              logFile.name = list[0]
+              return fs.readFile(path.join(tempFolder, logFile.name))
+            })
+            .then((contents) => {
+              logFile.lines = contents
+                .toString()
+                .split('\n')
+                .map(line => line.trim()) // remove any spaces, tabs, or \r chars
+                .filter(line => line.length > 0) // ignore empty lines
+            })
+        })
+
+        it('must end with the terminating signal', () => {
+          const pattern = /^exit code:\s+sigterm$/i
           const actual = logFile.lines[logFile.lines.length - 1]
 
           expect(actual).to.match(pattern)